Starting the changes

+++ Octokit/Clients/DeploymentStatusClient.cs
@@ -28,7 +28,7 @@ namespace Octokit
         /// <param name="name">The name of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
         [ManualRoute("GET", "/repos/{owner}/{repo}/deployments/{deployment_id}/statuses")]
-        public Task<IReadOnlyList<DeploymentStatus>> GetAll(string owner, string name, int deploymentId)
+        public Task<IReadOnlyList<DeploymentStatus>> GetAll(string owner, string name, long deploymentId)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -46,7 +46,7 @@ namespace Octokit
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
         [ManualRoute("GET", "/repositories/{id}/deployments/{deployment_id}/statuses")]
-        public Task<IReadOnlyList<DeploymentStatus>> GetAll(long repositoryId, int deploymentId)
+        public Task<IReadOnlyList<DeploymentStatus>> GetAll(long repositoryId, long deploymentId)
         {
             return GetAll(repositoryId, deploymentId, ApiOptions.None);
         }
@@ -65,7 +65,7 @@ namespace Octokit
         [Preview("ant-man")]
         [Preview("flash")]
         [ManualRoute("GET", "/repos/{owner}/{repo}/deployments/{deployment_id}/statuses")]
-        public Task<IReadOnlyList<DeploymentStatus>> GetAll(string owner, string name, int deploymentId, ApiOptions options)
+        public Task<IReadOnlyList<DeploymentStatus>> GetAll(string owner, string name, long deploymentId, ApiOptions options)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -90,7 +90,7 @@ namespace Octokit
         [Preview("ant-man")]
         [Preview("flash")]
         [ManualRoute("GET", "/repositories/{id}/deployments/{deployment_id}/statuses")]
-        public Task<IReadOnlyList<DeploymentStatus>> GetAll(long repositoryId, int deploymentId, ApiOptions options)
+        public Task<IReadOnlyList<DeploymentStatus>> GetAll(long repositoryId, long deploymentId, ApiOptions options)
         {
             Ensure.ArgumentNotNull(options, nameof(options));
 
@@ -114,7 +114,7 @@ namespace Octokit
         [Preview("ant-man")]
         [Preview("flash")]
         [ManualRoute("POST", "/repos/{owner}/{repo}/deployments/{deployment_id}/statuses")]
-        public Task<DeploymentStatus> Create(string owner, string name, int deploymentId, NewDeploymentStatus newDeploymentStatus)
+        public Task<DeploymentStatus> Create(string owner, string name, long deploymentId, NewDeploymentStatus newDeploymentStatus)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -138,7 +138,7 @@ namespace Octokit
         [Preview("ant-man")]
         [Preview("flash")]
         [ManualRoute("POST", "/repositories/{id}/deployments/{deployment_id}/statuses")]
-        public Task<DeploymentStatus> Create(long repositoryId, int deploymentId, NewDeploymentStatus newDeploymentStatus)
+        public Task<DeploymentStatus> Create(long repositoryId, long deploymentId, NewDeploymentStatus newDeploymentStatus)
         {
             Ensure.ArgumentNotNull(newDeploymentStatus, nameof(newDeploymentStatus));
